Move USD price calculation to Token Price Provider
