+++ balancer-js/src/modules/data-providers/token-price/static.provider.ts
@@ -1,8 +1,46 @@
+import { BigNumber, parseFixed } from '@ethersproject/bignumber';
 import { TokenPriceData } from '@/types';
 import { TokenPriceProvider } from './provider.interface';
 
 export class StaticTokenPriceProvider implements TokenPriceProvider {
-    constructor(private tokenPrices: TokenPriceData[]) {}
+    constructor(private tokenPrices: TokenPriceData[]) {
+        this.calculateUSDPrices();
+    }
+
+    /**
+     * Iterates through all tokens and calculates USD prices
+     * based on data the tokens already have.
+     */
+    calculateUSDPrices(): void {
+        const USDAssets = [
+            '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
+            '0x6b175474e89094c44da98b954eedeac495271d0f',
+            '0xdac17f958d2ee523a2206206994597c13d831ec7',
+        ];
+        let assetsAvailable = 0;
+        let assetValueSum = BigNumber.from(0);
+
+        USDAssets.forEach((address) => {
+            const tokenPrice = this.find(address);
+            if (tokenPrice?.ofNativeAsset) {
+                assetValueSum = assetValueSum.add(tokenPrice.ofNativeAsset);
+                assetsAvailable++;
+            }
+        });
+
+        const NativeAssetUSDPrice = assetValueSum.div(assetsAvailable);
+
+        this.tokenPrices = this.tokenPrices.map((tokenPrice) => {
+            if (tokenPrice.ofNativeAsset) {
+                tokenPrice.inUSD = parseFixed('1', 18)
+                    .mul(NativeAssetUSDPrice)
+                    .div(parseFixed(tokenPrice.ofNativeAsset, 18))
+                    .toString();
+            }
+
+            return tokenPrice;
+        });
+    }
 
     find(address: string): TokenPriceData | undefined {
         return this.tokenPrices.find((tokenPrice) => {
